move Opt to lib
